Create CollapsibleMixin.
Move src/CollapsableMixin => src/CollapsibleMixin
Create src/CollapsableMixin extended from src/CollapsibleMixin

Refactor CollapsibleMixin.dimension() method,
because it is the easiest way to get this method working.
The cause is that function used from component is optional.
I had tried `get` and `set`, but without success.

Copy tests for the new src/CollapsibleMixin.
Add assertions for deprecation warnings.

DRY tests.
